# api/src/openapi/apiSchema.json
{"openapi":"3.0.0","paths":{"/api/auth":{"post":{"operationId":"AuthController_getJWT","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TokenResult"}}}}}}},"/api/station":{"get":{"operationId":"StationController_getList","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/StationDto"}}}}},"401":{"description":""}},"security":[{"bearer":[]}]}},"/api/ticket":{"post":{"operationId":"TicketController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TicketCreateDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TicketDto"}}}},"400":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorDto"}}}},"401":{"description":""}},"security":[{"bearer":[]}]},"get":{"operationId":"TicketController_getList","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/TicketDto"}}}}}},"security":[{"bearer":[]}]}},"/api/ticket/{id}":{"patch":{"operationId":"TicketController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TicketUpdateDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TicketDto"}}}},"401":{"description":""}},"security":[{"bearer":[]}]},"get":{"operationId":"TicketController_get","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TicketDto"}}}}},"security":[{"bearer":[]}]}},"/api/train":{"get":{"operationId":"TrainController_getTrain","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TrainDto"}}}},"401":{"description":""}},"security":[{"bearer":[]}]}}},"info":{"title":"Spirit Pass API","description":"","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"TokenResult":{"type":"object","properties":{"token":{"type":"string"}},"required":["token"]},"StationDto":{"type":"object","properties":{"id":{"type":"number"},"name":{"type":"string"},"image":{"type":"string"}},"required":["id","name","image"]},"TicketDto":{"type":"object","properties":{"id":{"type":"number"},"trainId":{"type":"number"},"station":{"$ref":"#/components/schemas/StationDto"},"name":{"type":"string"},"issuedAt":{"format":"date-time","type":"string"}},"required":["id","trainId","station","name","issuedAt"]},"ErrorDto":{"type":"object","properties":{"error":{"type":"string"},"message":{"type":"string"},"statusCode":{"type":"number"}},"required":["error","message","statusCode"]},"TicketCreateDto":{"type":"object","properties":{"trainId":{"type":"number"},"stationId":{"type":"number"},"name":{"type":"string"}},"required":["trainId","stationId","name"]},"TicketUpdateDto":{"type":"object","properties":{"trainId":{"type":"number"},"stationId":{"type":"number"},"name":{"type":"string"}}},"TrainDto":{"type":"object","properties":{"id":{"type":"number"},"seats":{"type":"number"},"departureAt":{"format":"date-time","type":"string"},"availableTickets":{"type":"number"}},"required":["id","seats","departureAt","availableTickets"]}}}}